refactor

+++ BankingSystem/Controllers/AccountsController.cs
@@ -20,21 +20,18 @@ namespace BankingSystem.Controllers
             IMapper mapper
             )
         {
-            _accountService = accountService;
-            _mapper = mapper;
-            _userService = userService;
-
+            _accountService = accountService ?? throw new ArgumentNullException(nameof(accountService));
+            _mapper = mapper ?? throw new ArgumentNullException(nameof(mapper));
+            _userService = userService ?? throw new ArgumentNullException(nameof(userService));
         }
 
-        [HttpGet("GetAccount/{userId}/{accountId}", Name = "GetAccount")]
-        public async Task<ActionResult<Account>> GetAccount(int userId, int accountId)
+        [HttpGet("GetAccount/{accountId}", Name = "GetAccount")]
+        public async Task<ActionResult<Account>> GetAccount(int accountId)
         {
-            var user = await _userService.GetUser(userId);
+            if (accountId <= 0)
+                return BadRequest("Account ID must be greater than zero.");
 
-            if (user == null)
-                return NotFound();
-
-            var account = await _accountService.GetAccount(accountId);
+            var account = await _accountService.GetAccountAsync(accountId);
 
             if (account == null)
                 return NotFound();
@@ -46,40 +43,49 @@ namespace BankingSystem.Controllers
         [HttpGet("GetAccountList/{userId}")]
         public async Task<ActionResult<List<Account>>> GetAccountList(int userId)
         {
-            var accountList = await _accountService.GetAccountListByUserId(userId);
+            if (userId <= 0)
+                return BadRequest("User ID must be greater than zero.");
+
+            var accountList = await _accountService.GetAccountsByUserIdAsync(userId);
 
             return Ok(_mapper.Map<IEnumerable<AccountReadDto>>(accountList));
         }
 
-        [HttpPost("CreateAccount/{id}")]
-        public async Task<ActionResult<Account>> CreateAccount(int id)
+        [HttpPost("CreateAccount/{userId}")]
+        public async Task<ActionResult<Account>> CreateAccount(int userId)
         {
-            var user = await _userService.GetUser(id);
+            if (userId <= 0)
+                return BadRequest("User ID must be greater than zero.");
+
+            var user = await _userService.GetUser(userId);
 
             if (user == null)
                 return NotFound();
 
             var account = _accountService.CreateAccount(user);
 
-            await _accountService.SaveChange();
+            await _accountService.SaveChangesAsync();
             var result = _mapper.Map<AccountReadDto>(account);
 
-            return CreatedAtRoute(nameof(GetAccount), new { UserId = id, accountId = account.Id }, result);
+            return CreatedAtRoute(nameof(GetAccount), new { UserId = userId, accountId = account.Id }, result);
         }
 
 
         [HttpDelete("DeleteAccount/{accountId}")]
         public async Task<IActionResult> DeleteAccount(int accountId)
         {
-            var account = await _accountService.GetAccount(accountId);
+            if (accountId <= 0)
+                return BadRequest("Account ID must be greater than zero.");
+
+            var account = await _accountService.GetAccountAsync(accountId);
 
             if (account == null)
-                return NotFound("Account not found!");
+                return NotFound();
 
             try
             {
                 _accountService.DeleteAccount(account);
-                await _accountService.SaveChange();
+                await _accountService.SaveChangesAsync();
             }
             catch (Exception ex)
             {
@@ -97,7 +103,7 @@ namespace BankingSystem.Controllers
             if (user == null)
                 return NotFound("User not found!");
 
-            var account = await _accountService.GetAccount(accountDeposit.AccountId);
+            var account = await _accountService.GetAccountAsync(accountDeposit.AccountId);
 
             if (account == null)
                 return NotFound("Account not found!");
@@ -107,8 +113,8 @@ namespace BankingSystem.Controllers
 
             try
             {
-                _accountService.MoneyTransfer(account, accountDeposit);
-                await _accountService.SaveChange();
+                _accountService.TransferMoney(account, accountDeposit);
+                await _accountService.SaveChangesAsync();
             }
             catch (Exception ex)
             {
